fix time by changing to long

+++ BubbleSort.java
@@ -55,7 +55,7 @@ public class BubbleSort {
         return data;
     }
 
-    public int getTimeElapsed() {
+    public long getTimeElapsed() {
         return timeElapsed.getNano();
     }
