+++ InsertionSort.java
@@ -59,7 +59,7 @@ public class InsertionSort {
         return data;
     }
 
-    public int getTimeElapsed() {
+    public long getTimeElapsed() {
         return timeElapsed.getNano();
     }
