+++ SelectionSort.java
@@ -65,7 +65,7 @@ public class SelectionSort {
         return data;
     }
 
-    public int getTimeElapsed() {
+    public long getTimeElapsed() {
         return timeElapsed.getNano();
     }
